[Next Gen] refactor(codegen): split v-bind prop emission into vbind
