+++ crates/vize_atelier_core/src/codegen/props.rs
@@ -5,6 +5,7 @@ mod events;
 mod generate;
 mod object;
 mod scan;
+mod vbind;
 mod von_vmodel;
 
 use crate::{PropNode, RuntimeHelper};
